requirements.txt
streamlit
openai
beautifulsoup4
requests
python-dateutil
psycopg2
sqlite3 